# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dasmesh Stock Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opening-stock:
    get:
      operationId: listOpeningStock
      tags: [openingStock]
      summary: List opening stock balances
      responses:
        "200":
          description: Opening stock list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpeningStockEntry"
    post:
      operationId: setOpeningStock
      tags: [openingStock]
      summary: Set opening stock for an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpeningStockInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpeningStockEntry"
  /opening-stock/import:
    post:
      operationId: importOpeningStock
      tags: [openingStock]
      summary: Import opening stock from CSV/Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpeningStockImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # PRODUCTION ENTRIES
...
  /dispatch:
    get:
      operationId: listDispatch
      tags: [dispatch]
      summary: List dispatch entries
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
        - name: stockItemId
          in: query
          schema:
            type: integer
        - name: partyName
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dispatch entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DispatchEntry"
    post:
      operationId: createDispatch
      tags: [dispatch]
      summary: Create a dispatch entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchEntry"
  /dispatch/{id}:
    get:
      operationId: getDispatch
      tags: [dispatch]
      summary: Get a dispatch entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatch entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchEntry"
    patch:
      operationId: updateDispatch
      tags: [dispatch]
      summary: Update a dispatch entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchEntry"
    delete:
      operationId: deleteDispatch
      tags: [dispatch]
      summary: Delete a dispatch entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # STOCK REGISTER
...
components:
  schemas:
...
    StockItem:
      type: object
      required: [id, itemCode, category, size, length, unit, status]
      properties:
        id:
          type: integer
        itemCode:
          type: string
        category:
          type: string
        size:
          type: string
        length:
          type: string
        unit:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
...
    OpeningStockEntry:
      type: object
      required: [id, stockItemId, quantity, effectiveDate]
      properties:
        id:
          type: integer
        stockItemId:
          type: integer
        stockItem:
          $ref: "#/components/schemas/StockItem"
        quantity:
          type: number
        effectiveDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    OpeningStockInput:
      type: object
      required: [stockItemId, quantity, effectiveDate]
      properties:
        stockItemId:
          type: integer
        quantity:
          type: number
        effectiveDate:
          type: string
          format: date
    OpeningStockImport:
      type: object
      required: [effectiveDate, rows]
      properties:
        effectiveDate:
          type: string
          format: date
        rows:
          type: array
          items:
            type: object
            required: [itemCode, quantity]
            properties:
              itemCode:
                type: string
              quantity:
                type: number
    ImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string

    # PRODUCTION
...
    DispatchEntry:
      type: object
      required: [id, date, stockItemId, quantity, partyName]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        stockItemId:
          type: integer
        stockItem:
          $ref: "#/components/schemas/StockItem"
        quantity:
          type: number
        partyName:
          type: string
        invoiceNumber:
          type: ["string", "null"]
        vehicleNumber:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: ["string", "null"]
        updatedById:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    DispatchInput:
      type: object
      required: [date, stockItemId, quantity, partyName]
      properties:
        date:
          type: string
          format: date
        stockItemId:
          type: integer
        quantity:
          type: number
        partyName:
          type: string
        invoiceNumber:
          type: string
        vehicleNumber:
          type: string
        remarks:
          type: string
    DispatchUpdate:
      type: object
      properties:
        quantity:
          type: number
        partyName:
          type: string
        invoiceNumber:
          type: string
        vehicleNumber:
          type: string
        remarks:
          type: string

    # STOCK REGISTER
...