# lib/api-zod/src/generated/types/openingStockImport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dasmesh Stock Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OpeningStockImportRowsItem } from './openingStockImportRowsItem';

export interface OpeningStockImport {
  effectiveDate: Date;
  rows: OpeningStockImportRowsItem[];
}